[NAE-1061] Resource pagination with paginator footer
- reduce the duplication code
